refactor(defaults): share sys default helpers with web bundle
