+++ frappe/public/js/frappe/defaults.js
@@ -1,7 +1,9 @@
 // Copyright (c) 2015, Frappe Technologies Pvt. Ltd. and Contributors
 // MIT License. See license.txt
 
-frappe.defaults = {
+import "./sys_defaults.js";
+
+Object.assign(frappe.defaults, {
 	get_user_default: function (key) {
 		let defaults = frappe.boot.user.defaults;
 		let d = defaults[key];
@@ -57,19 +59,6 @@ frappe.defaults = {
 		// filter out values which are not permitted to the user
 		d = d.filter((item) => frappe.defaults.in_user_permission(key, item));
 		return d;
-	},
-	get_global_default: function (key) {
-		var d = frappe.sys_defaults[key];
-		if ($.isArray(d)) d = d[0];
-		return d;
-	},
-	get_global_defaults: function (key) {
-		var d = frappe.sys_defaults[key];
-		if (!$.isArray(d)) d = [d];
-		return d;
-	},
-	is_enabled: function (key) {
-		return cint(this.get_global_default(key)) === 1;
 	},
 	set_user_default_local: function (key, value) {
 		frappe.boot.user.defaults[key] = value;
@@ -137,4 +126,4 @@ frappe.defaults = {
 			frappe.defaults.update_user_permissions();
 		}
 	},
-};
+});
